Generate SSH keys during make `docker-init`
The SSH keys were being generated with a fixed secret key (which is not
ideal for security reasons, although given the scope of these files and
the fact that they are meant to be used in development environments
means that no actual security concern was real). The keys are now
generated using a random key, which is stored in the local, non
committable, .env file.