Ensure the public directory is reachable
The `nginx` container previously pointed the root of the site to the
hardcoded /var/www/html directory. However, the installation script
already places the test files in a user-specified public directory. This
commit uses that folder as the root of the `nginx` site by changing the
`default.conf` configuration file on installation.