Allow the www-data user to use nvm, node and friends
This lead to a rewrite of many installation steps for node and composer,
specifically because we were hard-coding the full ~/.bash_profile file
into the container. We now append to what was there. At the moment, this
is irrelevant, because the image starts *without* that file and we could
simply add what we want. But if this image grown into something more,
these changes ensure that no overwriting step occurrs that could
ultimately destroy previous changes.